refactor: make EmptySearchLayout fill the screen size
